requirements.txt
numpy
pillow
matplotlib
tqdm
tensorflow
keras
